@@ -11,12 +12,13 @@ if (localPropertiesFile.exists()) {
 }
 
 android {
-    compileSdkVersion 31
+    compileSdkVersion 32
 
     defaultConfig {
         applicationId "info.firozansari.movieapp"
+        targetSdkVersion 32
         minSdkVersion 26
-        targetSdkVersion 31
+
         versionCode 2
         versionName "1.1"
         buildConfigField("String","TMDB_API_KEY", localProperties.getProperty("tmdb_api_key") ?: "\"\"")
